finalizacion del sistema
se pueden realizar ventas y guadar sus detalles en la base de datos

+++ RafaelGonzalez/sistemas_Ventas/Vista/frmUsuarios.cs
@@ -20,6 +20,7 @@ namespace sistemas_Ventas.Vista
             InitializeComponent();
         }
          Usuarios user = new Usuarios();
+        DialogResult R;
         void Limpiartext()
         {
             txtEmail.Text = "";
@@ -27,7 +28,7 @@ namespace sistemas_Ventas.Vista
         }
         void cargardatos()
         {
-            using (sistema_ventasEntities2 db = new sistema_ventasEntities2())
+            using (sistema_ventasEntities4 db = new sistema_ventasEntities4())
             {
                 var lista = from u in db.Usuarios
                             select new { u.Id,u.Email,u.Contrasena};
@@ -43,28 +44,30 @@ namespace sistemas_Ventas.Vista
         private void Usuarios_Load(object sender, EventArgs e)
         {
             cargardatos();
+            DesactivarbotonesEditarEliminar();
+            btnGuardar.Enabled = false;
+        }
+        public void activarTex()
+        {
+            txtContraseña.Enabled = true;
+            txtEmail.Enabled = true;
         }
 
-        private void btnCargar_Click(object sender, EventArgs e)
-        {
-            
-            
-            
-        }
+      
 
         private void btnGuardar_Click(object sender, EventArgs e)
         {
-            using (sistema_ventasEntities2 db = new sistema_ventasEntities2())
+            using (sistema_ventasEntities4 db = new sistema_ventasEntities4())
             {
                
+                    user.Contrasena = txtContraseña.Text;
+                    user.Email = txtEmail.Text;
 
-                user.Contrasena = txtContraseña.Text;
-                user.Email = txtEmail.Text;
-
-                db.Usuarios.Add(user);
-                db.SaveChanges();
-                Limpiartext();
-                cargardatos();
+                    db.Usuarios.Add(user);
+                    db.SaveChanges();
+                    Limpiartext();
+                    cargardatos();
+               
 
             }
         }
@@ -72,9 +75,12 @@ namespace sistemas_Ventas.Vista
         private void btnEditar_Click(object sender, EventArgs e)
         {
            
-            using (sistema_ventasEntities2 db = new sistema_ventasEntities2())
+            using (sistema_ventasEntities4 db = new sistema_ventasEntities4())
             {
-                
+            R = MessageBox.Show("Desea guardar los cambios", "Alerta", MessageBoxButtons.YesNo, MessageBoxIcon.Information, MessageBoxDefaultButton.Button2);
+            if (R == DialogResult.Yes)
+            {
+
                 string Id = dtvUsuarios.CurrentRow.Cells[0].Value.ToString();
                 int Idc = int.Parse(Id);
                 user = db.Usuarios.Where(verificarId => verificarId.Id == Idc).First();
@@ -84,19 +90,39 @@ namespace sistemas_Ventas.Vista
                 db.SaveChanges();
                 Limpiartext();
                 cargardatos();
+                }
+                else
+                {
+                    Limpiartext();
+                    cargardatos();
+                }
+                DesactivarbotonesEditarEliminar();
             }
+
          }
 
         private void btnEliminar_Click(object sender, EventArgs e)
         {
             string Id = dtvUsuarios.CurrentRow.Cells[0].Value.ToString();
-            using (sistema_ventasEntities2 db = new sistema_ventasEntities2())
+            using (sistema_ventasEntities4 db = new sistema_ventasEntities4())
             {
-                user = db.Usuarios.Find(int.Parse(Id));
-                db.Usuarios.Remove(user);
-                db.SaveChanges();
-                Limpiartext();
-                cargardatos();
+                R = MessageBox.Show("Desea eliminar el registro", "Alerta", MessageBoxButtons.YesNo, MessageBoxIcon.Information, MessageBoxDefaultButton.Button2);
+                if (R == DialogResult.Yes)
+                {
+                    user = db.Usuarios.Find(int.Parse(Id));
+                    db.Usuarios.Remove(user);
+                    db.SaveChanges();
+                    Limpiartext();
+                    cargardatos();
+                }
+                else
+                {
+                    Limpiartext();
+                    cargardatos();
+
+                }
+                DesactivarbotonesEditarEliminar();
+                
             }
 
         }
@@ -108,6 +134,30 @@ namespace sistemas_Ventas.Vista
 
             txtContraseña.Text = Contra;
             txtEmail.Text = Email;
+            activarBotones();
+        }
+        void DesactivarbotonesEditarEliminar()
+        {
+            btnEditar.Enabled = false;
+            btnEliminar.Enabled = false;
+            btnGuardar.Enabled = true;
+            txtContraseña.Enabled = false;
+            txtEmail.Enabled = false;
+        }
+
+        void activarBotones()
+        {
+            btnGuardar.Enabled = false;
+            btnEditar.Enabled = true;
+            btnEliminar.Enabled = true;
+        }
+
+        private void button1_Click(object sender, EventArgs e)
+        {
+            btnGuardar.Enabled = true;
+            DesactivarbotonesEditarEliminar();
+            activarTex();
+            
         }
     }
 }
